+++ R/calculate_diff_abundance.R
@@ -60,16 +60,17 @@ diff_abundance <-
 #' Is not required if \code{method = "t-test_mean_sd"}.
 #' @param condition a character or numeric column in the \code{data} data frame that contains the
 #' conditions.
-#' @param grouping a character column in the \code{data} data frame that contains precursor or
-#' peptide identifiers.
+#' @param grouping a character column in the \code{data} data frame that contains precursor, 
+#' peptide or protein identifiers.
 #' @param intensity_log2 a numeric column in the \code{data} data frame that contains intensity
 #' values. The intensity values need to be log2 transformed. Is not required if
 #' \code{method = "t-test_mean_sd"}.
 #' @param missingness a character column in the \code{data} data frame that contains missingness
 #' information. Can be obtained by calling \code{assign_missingness()}. Is not required if
 #' \code{method = "t-test_mean_sd"}. The type of missingness assigned to a comparison does not have
-#' any influence on the statistical test. However, if \code{filter_NA_missingness = TRUE} then
-#' comparisons with missingness \code{NA} are filtered out prior to p-value adjustment.
+#' any influence on the statistical test. However, if \code{filter_NA_missingness = TRUE} and 
+#' \code{method = "proDA"}, then comparisons with missingness \code{NA} are filtered out prior 
+#' to p-value adjustment.
 #' @param comparison a character column in the \code{data} data frame that contains information of
 #' treatment/reference condition pairs. Can be obtained by calling \code{assign_missingness}.
 #' Comparisons need to be in the form condition1_vs_condition2, meaning two compared conditions are
@@ -91,8 +92,8 @@ diff_abundance <-
 #' for example obtained by calling \code{assign_missingness()}. If a reference/treatment pair has
 #' too few samples to be considered robust based on user defined cutoffs, it is annotated with \code{NA}
 #' as missingness by the \code{assign_missingness()} function. If this argument is \code{TRUE},
-#' these \code{NA} reference/treatment pairs are filtered out after the testing and prior to p-value
-#' adjustment.
+#' these \code{NA} reference/treatment pairs are filtered out. For \code{method = "proDA"} this
+#' is done before the p-value adjustment.
 #' @param method a character value, specifies the method used for statistical hypothesis testing.
 #' Methods include Welch test (\code{"t-test"}), a Welch test on means, standard deviations and
 #' number of replicates ("\code{"t-test_mean_sd"}) and a moderated t-test based on the \code{limma}
@@ -109,7 +110,8 @@ diff_abundance <-
 #' @param retain_columns a vector indicating if certain columns should be retained from the input
 #' data frame. Default is not retaining additional columns \code{retain_columns = NULL}. Specific
 #' columns can be retained by providing their names (not in quotations marks, just like other
-#' column names, but in a vector).
+#' column names, but in a vector). Please note that if you retain columns that have multiple 
+#' rows per grouped variable there will be duplicated rows in the output.
 #'
 #' @return A data frame that contains differential abundances (\code{diff}), p-values (\code{pval})
 #' and adjusted p-values (\code{adj_pval}) for each protein, peptide or precursor (depending on
@@ -144,6 +146,10 @@ diff_abundance <-
 #' \code{n_obs} contains the number of observations for the specific protein, peptide or precursor
 #' (depending on the \code{grouping} variable) and the associated treatment/reference pair.}
 #' }
+#' For all methods execept \code{"proDA"}, the p-value adjustment is performed only on the 
+#' proportion of data that contains a p-value that is not \code{NA}. For \code{"proDA"} the
+#' p-value adjustment is either performed on the complete dataset (\code{filter_NA_missingness = TRUE})
+#' or on the subset of the dataset with missingness that is not \code{NA} (\code{filter_NA_missingness = FALSE}).
 #' @import dplyr
 #' @import tidyr
 #' @importFrom rlang .data enquo ensym as_name as_label expr := !!
@@ -297,22 +303,31 @@ Please provide a valid reference condition.", prefix = "\n", initial = ""))
           }
         )) %>%
         dplyr::mutate(diff = ifelse(diff == "NaN", NA, diff)) %>%
+        dplyr::select(-c(.data$control, .data$treated)) %>%
+        dplyr::left_join(t_test_missingness_obs, by = c(rlang::as_name(rlang::enquo(grouping)), "comparison")) 
+      
+      # save all p-values that are NA to join back after adjustment. Otherwise adjustment is wrong since 
+      # the p.adjust function does not remove NA values
+      t_test_NA_pval <- t_test_result %>% 
+        dplyr::filter(is.na(.data$pval))
+      
+      # Adjust the p-values for multiple testing
+      t_test_adj_pval <- t_test_result %>% 
+        dplyr::filter(!is.na(.data$pval)) %>% 
         dplyr::group_by({{ comparison }}) %>%
         dplyr::mutate(adj_pval = stats::p.adjust(.data$pval, method = p_adj_method)) %>%
         dplyr::ungroup() %>%
-        dplyr::select(-c(.data$control, .data$treated)) %>%
-        dplyr::left_join(t_test_missingness_obs, by = c(rlang::as_name(rlang::enquo(grouping)), "comparison")) %>%
+        dplyr::bind_rows(t_test_NA_pval) %>% 
         dplyr::arrange(.data$adj_pval, .data$pval)
 
       message("DONE", appendLF = TRUE)
 
       if (!missing(retain_columns)) {
-        t_test_result <- data %>%
+        t_test_adj_pval <- data %>%
           dplyr::ungroup() %>%
           dplyr::select(
             !!enquo(retain_columns),
-            {{ intensity_log2 }},
-            colnames(t_test_result)[!colnames(t_test_result) %in%
+            colnames(t_test_adj_pval)[!colnames(t_test_adj_pval) %in%
               c(
                 "pval",
                 "std_error",
@@ -321,10 +336,8 @@ Please provide a valid reference condition.", prefix = "\n", initial = ""))
                 "n_obs"
               )]
           ) %>%
-          tidyr::drop_na({{ intensity_log2 }}) %>%
-          dplyr::select(-{{ intensity_log2 }}) %>%
           dplyr::distinct() %>%
-          dplyr::right_join(t_test_result, by = colnames(t_test_result)[!colnames(t_test_result) %in%
+          dplyr::right_join(t_test_adj_pval, by = colnames(t_test_adj_pval)[!colnames(t_test_adj_pval) %in%
             c(
               "pval",
               "std_error",
@@ -336,16 +349,12 @@ Please provide a valid reference condition.", prefix = "\n", initial = ""))
       }
 
       if (filter_NA_missingness == TRUE) {
-        t_test_result <- t_test_result %>%
-          tidyr::drop_na({{ missingness }}) %>%
-          dplyr::group_by({{ comparison }}) %>%
-          dplyr::mutate(adj_pval = stats::p.adjust(.data$pval, method = p_adj_method)) %>%
-          dplyr::ungroup() %>%
-          dplyr::arrange(.data$adj_pval, .data$pval)
-        return(t_test_result)
+        t_test_adj_pval <- t_test_adj_pval %>%
+          tidyr::drop_na({{ missingness }})
+        return(t_test_adj_pval)
       }
       if (filter_NA_missingness == FALSE) {
-        return(t_test_result)
+        return(t_test_adj_pval)
       }
     }
 
@@ -578,16 +587,29 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
           adj_pval = .data$adj.P.Val
         ) %>%
         dplyr::left_join(moderated_t_test_missingness, by = c(rlang::as_name(rlang::enquo(grouping)), "comparison"))
+      
+      # save all p-values that are NA to join back after adjustment. Otherwise adjustment is wrong since 
+      # the p.adjust function does not remove NA values
+      moderated_t_test_NA_pval <- moderated_t_test_result %>% 
+        dplyr::filter(is.na(.data$pval))
+      
+      # Adjust the p-values for multiple testing
+      moderated_t_test_adj_pval <- moderated_t_test_result %>% 
+        dplyr::filter(!is.na(.data$pval)) %>% 
+        dplyr::group_by({{ comparison }}) %>%
+        dplyr::mutate(adj_pval = stats::p.adjust(.data$pval, method = p_adj_method)) %>%
+        dplyr::ungroup() %>%
+        dplyr::bind_rows(moderated_t_test_NA_pval) %>% 
+        dplyr::arrange(.data$adj_pval, .data$pval)
 
       message("DONE", appendLF = TRUE)
 
       if (!missing(retain_columns)) {
-        moderated_t_test_result <- data %>%
+        moderated_t_test_adj_pval <- data %>%
           dplyr::ungroup() %>%
           dplyr::select(
             !!enquo(retain_columns),
-            {{ intensity_log2 }},
-            colnames(moderated_t_test_result)[!colnames(moderated_t_test_result) %in%
+            colnames(moderated_t_test_adj_pval)[!colnames(moderated_t_test_adj_pval) %in%
               c(
                 "CI_2.5",
                 "CI_97.5",
@@ -600,10 +622,8 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
                 "n_obs"
               )]
           ) %>%
-          tidyr::drop_na({{ intensity_log2 }}) %>%
-          dplyr::select(-{{ intensity_log2 }}) %>%
           dplyr::distinct() %>%
-          dplyr::right_join(moderated_t_test_result, by = colnames(moderated_t_test_result)[!colnames(moderated_t_test_result) %in%
+          dplyr::right_join(moderated_t_test_adj_pval, by = colnames(moderated_t_test_adj_pval)[!colnames(moderated_t_test_adj_pval) %in%
             c(
               "CI_2.5",
               "CI_97.5",
@@ -619,16 +639,12 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
       }
 
       if (filter_NA_missingness == TRUE) {
-        moderated_t_test_result <- moderated_t_test_result %>%
-          tidyr::drop_na({{ missingness }}) %>%
-          dplyr::group_by(.data$comparison) %>%
-          dplyr::mutate(adj_pval = stats::p.adjust(.data$pval, method = p_adj_method)) %>%
-          dplyr::ungroup() %>%
-          dplyr::arrange(.data$adj_pval, .data$pval)
-        return(moderated_t_test_result)
+        moderated_t_test_adj_pval <- moderated_t_test_adj_pval %>%
+          tidyr::drop_na({{ missingness }}) 
+        return(moderated_t_test_adj_pval)
       }
       if (filter_NA_missingness == FALSE) {
-        return(moderated_t_test_result)
+        return(moderated_t_test_adj_pval)
       }
     }
 
@@ -666,15 +682,15 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
       message("DONE", appendLF = TRUE)
       message("[4/5] Define missingness levels for filtering ... ", appendLF = FALSE)
 
-      proDA_missingness <-
-        data %>%
-        tidyr::drop_na({{ missingness }}, {{ intensity_log2 }}) %>%
+      proDA_missingness <- data %>%
+        tidyr::drop_na({{ intensity_log2 }}) %>%
         dplyr::group_by({{ comparison }}, {{ grouping }}) %>%
         dplyr::mutate(n_obs = dplyr::n()) %>%
         dplyr::ungroup() %>%
         dplyr::distinct({{ grouping }}, {{ comparison }}, {{ missingness }}, .data$n_obs)
 
       proDA_filter <- proDA_missingness %>%
+        tidyr::drop_na({{ missingness }}) %>% 
         dplyr::distinct({{ grouping }}, {{ comparison }}) %>%
         split(.$comparison) %>%
         purrr::map(dplyr::select, -{{ comparison }})
@@ -741,7 +757,6 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
             .f = ~ dplyr::mutate(.x, comparison = str_replace_all(.y, pattern = "`", replacement = ""))
           ) %>%
           purrr::map_dfr(~ dplyr::mutate(.x, adj_pval = p.adjust(.data$pval, method = p_adj_method))) %>%
-          dplyr::select(-.data$n_obs) %>%
           dplyr::select(-.data$n_obs, -.data$n_approx) %>%
           dplyr::rename({{ grouping }} := .data$name, std_error = .data$se) %>%
           dplyr::left_join(proDA_missingness, by = c(rlang::as_name(rlang::enquo(grouping)), "comparison"))
@@ -754,7 +769,6 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
           dplyr::ungroup() %>%
           dplyr::select(
             !!enquo(retain_columns),
-            {{ intensity_log2 }},
             colnames(proDA_result)[!colnames(proDA_result) %in%
               c(
                 "std_error",
@@ -767,8 +781,6 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
                 "n_obs"
               )]
           ) %>%
-          tidyr::drop_na({{ intensity_log2 }}) %>%
-          dplyr::select(-{{ intensity_log2 }}) %>%
           dplyr::distinct() %>%
           dplyr::right_join(proDA_result, by = colnames(proDA_result)[!colnames(proDA_result) %in%
             c(